@@ -10,9 +10,13 @@
   },
   "jest": {
     "preset": "jest-expo",
+    "setupFilesAfterEnv": [
+      "@testing-library/jest-native/extend-expect"
+    ],
     "verbose": true,
+    "resetMocks": true,
     "transformIgnorePatterns": [
-      "node_modules/(?!((jest-)?react-native(-.*)?|@react-native(-community)?|@react-navigation|@rneui?|expo?|react-native-vector-icons)/)"
+      "node_modules/(?!((jest-)?react-native(-.*)?|@react-native(-community)?|@react-navigation|@rneui?|@react-native-vector-icons|@firebase?|firebase?|expo-constants|expo-modules-core|@sentry/.*|expo-linear-gradient)/)"
     ],
     "transform": {
       "^.+\\.(js|ts|tsx)$": "babel-jest"
@@ -39,7 +43,7 @@
     "@expo-google-fonts/mulish": "^0.2.3",
     "@expo-google-fonts/oswald": "^0.2.3",
     "@firebase/rules-unit-testing": "^2.0.7",
-    "@react-native-community/datetimepicker": "6.7.3",
+    "@react-native-community/datetimepicker": "7.2.0",
     "@react-navigation/bottom-tabs": "^6.5.7",
     "@react-navigation/native": "^6.1.6",
     "@react-navigation/native-stack": "^6.9.12",
@@ -49,34 +53,32 @@
     "@testing-library/react-hooks": "^7.0.2",
     "axios": "^1.4.0",
     "dotenv": "^16.0.3",
-    "expo": "~48.0.18",
-    "expo-constants": "~14.2.1",
-    "expo-dev-client": "~2.2.1",
-    "expo-device": "~5.2.1",
-    "expo-font": "~11.1.1",
-    "expo-linear-gradient": "~12.1.2",
-    "expo-notifications": "~0.18.1",
-    "expo-splash-screen": "~0.18.2",
-    "expo-status-bar": "~1.4.4",
-    "expo-system-ui": "~2.2.1",
-    "expo-updates": "~0.16.4",
+    "expo": "^49.0.6",
+    "expo-constants": "~14.4.2",
+    "expo-dev-client": "~2.4.6",
+    "expo-device": "~5.4.0",
+    "expo-font": "~11.4.0",
+    "expo-linear-gradient": "~12.3.0",
+    "expo-notifications": "~0.20.1",
+    "expo-splash-screen": "~0.20.5",
+    "expo-status-bar": "~1.6.0",
+    "expo-system-ui": "~2.4.0",
+    "expo-updates": "~0.18.11",
     "firebase": "^9.21.0",
-    "jest": "^29.2.1",
-    "jest-expo": "^48.0.2",
     "mobx-react-lite": "^3.4.3",
     "openai": "^3.2.1",
     "react": "18.2.0",
-    "react-native": "0.71.8",
+    "react-native": "0.72.3",
     "react-native-flip-card": "^3.5.7",
     "react-native-form-component": "^2.6.10",
-    "react-native-gesture-handler": "~2.9.0",
+    "react-native-gesture-handler": "~2.12.0",
     "react-native-global-state-hooks": "^5.0.14",
-    "react-native-reanimated": "~2.14.4",
+    "react-native-reanimated": "~3.3.0",
     "react-native-reanimated-carousel": "^3.3.2",
-    "react-native-safe-area-context": "4.5.0",
-    "react-native-screens": "~3.20.0",
+    "react-native-safe-area-context": "4.6.3",
+    "react-native-screens": "~3.22.0",
     "react-native-vector-icons": "^9.2.0",
-    "react-native-web": "~0.18.10",
+    "react-native-web": "~0.19.6",
     "zodiac-signs": "^1.2.1"
   },
   "devDependencies": {
@@ -88,12 +90,14 @@
     "@testing-library/jest-native": "^5.4.2",
     "@testing-library/react-native": "^12.1.2",
     "@types/jest": "^29.5.2",
-    "@types/react": "~18.0.14",
+    "@types/react": "~18.2.14",
     "@types/react-native-flip-card": "^3.5.2",
     "@types/react-test-renderer": "^18.0.0",
     "babel-jest": "^29.5.0",
+    "jest": "^29.6.2",
+    "jest-expo": "^49.0.0",
     "ts-jest": "^29.1.0",
-    "typescript": "^4.9.4"
+    "typescript": "^5.1.3"
   },
   "private": true
 }
