@@ -1,6 +1,6 @@
-import { getAuth } from './__mocks__/firebaseConfig';
+import { getAuth } from './__mocks__/config/firebaseConfig';
 
-jest.mock('./__mocks__/firebaseConfig');
+jest.mock('./__mocks__/config/firebaseConfig');
 
 test('getAuth is mocked', () => {
   getAuth();
