@@ -1,15 +1,28 @@
 // __mocks__/firebaseConfig.js
 
-export const getAuth = jest.fn();
+export const getAuth = jest.fn(() => ({
+  currentUser: {
+    email: "test@example.com",
+    uid: "123456789",
+  },
+  signInWithEmailAndPassword: jest.fn(),
+  signOut: jest.fn(),
+}));
 export const createUserWithEmailAndPassword = jest.fn();
 export const signInWithEmailAndPassword = jest.fn();
 export const firestore = {
   collection: jest.fn(() => ({
     doc: jest.fn(() => ({
-      get: jest.fn(() => Promise.resolve({
-        exists: true,
-        data: jest.fn(() => ({ /* les données de votre document mocké ici */})),
-      })),
+      get: jest.fn(() =>
+        Promise.resolve({
+          exists: true,
+          data: jest.fn(() => ({
+            prenom: "John",
+            nom: "Doe",
+            email:"john.doe@example.com"
+          })),
+        })
+      ),
       set: jest.fn(() => Promise.resolve(true)),
     })),
   })),
